refactor: HealthManageCard 관련로직 리팩토링
