+++ src/assets/svgs/index.ts
@@ -1,13 +1,13 @@
 export { ReactComponent as LogoImage } from './logo.svg'
 
-export { ReactComponent as HealthIcon1 } from './healthManage/ic-icon-mission-h-1.svg'
-export { ReactComponent as HealthIcon2 } from './healthManage/ic-icon-mission-h-2.svg'
-export { ReactComponent as HealthIcon3 } from './healthManage/ic-icon-mission-h-3.svg'
-export { ReactComponent as HealthIcon4 } from './healthManage/ic-icon-mission-h-4.svg'
-export { ReactComponent as HealthIcon5 } from './healthManage/ic-icon-mission-h-5.svg'
-export { ReactComponent as HealthIcon6 } from './healthManage/ic-icon-mission-h-6.svg'
-export { ReactComponent as HealthIcon7 } from './healthManage/ic-icon-mission-h-7.svg'
-export { ReactComponent as HealthIcon8 } from './healthManage/ic-icon-mission-h-8.svg'
+export { ReactComponent as BloodPressureIcon } from './healthManage/ic-icon-mission-h-1.svg'
+export { ReactComponent as BMIIcon } from './healthManage/ic-icon-mission-h-2.svg'
+export { ReactComponent as DrnkIcon } from './healthManage/ic-icon-mission-h-3.svg'
+export { ReactComponent as GFRIcon } from './healthManage/ic-icon-mission-h-4.svg'
+export { ReactComponent as CholesterolIcon } from './healthManage/ic-icon-mission-h-5.svg'
+export { ReactComponent as SmokeIcon } from './healthManage/ic-icon-mission-h-6.svg'
+export { ReactComponent as FastingBloodSugerIcon } from './healthManage/ic-icon-mission-h-7.svg'
+export { ReactComponent as ExerciseIcon } from './healthManage/ic-icon-mission-h-8.svg'
 
 export { ReactComponent as HealthInfo } from './healthInfo/ic-icon-health-info.svg'
 export { ReactComponent as HealthInfoArrow } from './healthInfo/ic-icon-health-info-arrow.svg'
