+++ src/services/health.ts
@@ -65,7 +65,7 @@ export const fetchYearsChartInfo = () => {
   const scoreAndYears: YearsType[] = []
 
   value.forEach((score, i) => {
-    let obj: YearsType = {}
+    const obj: YearsType = {}
     obj.value = Number(score)
     obj.year = year[i]
     scoreAndYears.push(obj)
@@ -206,7 +206,7 @@ export const getHealthManageData = () => {
 
   const { boj, paramMap }: { boj: IHealthManageData; paramMap: IHealthManageData } = wxcResultMap
 
-  const healthMangeCardData = property.map((value) => {
+  const healthManageCardData = property.map((value) => {
     const tag: string[] = []
     healthTagList.forEach((currentValue) => {
       if (currentValue.tagId === value) tag.push(currentValue.tag1, currentValue.tag2, currentValue.tag3)
@@ -227,7 +227,7 @@ export const getHealthManageData = () => {
 
   return {
     wMymaxHscoreDy,
-    healthMangeCardData,
+    healthManageCardData,
   }
 }
