+++ src/types/health.d.ts
@@ -2,3 +2,10 @@ export interface YearsType {
   value?: number
   year?: string
 }
+
+export interface IHealthManageCardData {
+  title: string
+  value: string
+  boj: string[]
+  tag: string[]
+}
